fix as comments suggested
