+++ packages/services/testimonials.ts
@@ -6,6 +6,12 @@ import {
 } from '@packages/repositories/firestore';
 import { Testimonial } from '@packages/entities/testimonials';
 
+interface testimonialProps {
+  name: string;
+  testimonial: string;
+  gitUsername: string;
+}
+
 let testimonials: undefined | FirestoreDAO<Testimonial>;
 export function getTestimonialInstance() {
   if (testimonials != null) return testimonials;
@@ -22,4 +28,28 @@ function processQuestionSnapshot(doc: QueryDocumentSnapshot<DocumentData>) {
     ...doc.data(),
     id: doc.id,
   } as Testimonial;
+}
+
+export async function addTestimonial({
+  name,
+  testimonial,
+  gitUsername,
+}: testimonialProps) {
+  const testimonialsService = getTestimonialInstance();
+  try {
+    const member = await fetch(
+      `https://api.github.com/users/${gitUsername}`,
+    ).then((r) => r.json());
+
+    await testimonialsService.add({
+      name: name,
+      text: testimonial,
+      profileUrl: member.html_url,
+      avatarUrl: member.avatar_url,
+      approved: false,
+    });
+  } catch (e) {
+    return 1;
+  }
+  return 0;
 }
